consolidate artQuiz files. name scheme refactoring.
