+++ src/Apps/ArtQuiz/ArtQuizWelcome.tsx
