--- src/Components/ArtQuiz/Components/ArtQuizLeftPanel.tsx
@@ -1,5 +1,0 @@
-import { Text } from "@artsy/palette"
-
-export const ArtQuizLeftPanel = () => {
-  return <Text variant="xl">Art Quiz</Text>
-}
